refactor(course-modules): clean up unused imports
Removes a large number of unused imports from the `course-modules.tsx`
component, improving readability and reducing unnecessary dependencies.
