+++ frontend/src/pages/shared/courses/$courseId/grades/course-grades.tsx
@@ -1,30 +1,17 @@
-import type { IUsersQuery } from '@/features/user-management/types.ts'
-import {
-  Badge,
-  Box,
-  Button,
-  Center,
-  Flex,
-  Group,
-  rem,
-  Skeleton,
-  Stack,
-  Table,
-  Text,
-  Title,
-  Tooltip,
-} from '@mantine/core'
-import { Suspense, useState } from 'react'
-import { formatTimestampToDateTimeText } from '@/utils/formatters.ts'
-import SearchComponent from '@/components/search-component.tsx'
 import type {
   CourseGradebookForMentor,
   CourseGradebookForStudent,
-  Grade,
   StudentAssignmentGrade,
 } from '@/features/courses/grades/types.ts'
 import { useAuth } from '@/features/auth/auth.hook.ts'
-import { IconNote } from '@tabler/icons-react'
+import { useState } from 'react'
+import { Button, Group, rem } from '@mantine/core'
+import SearchComponent from '@/components/search-component.tsx'
+import {
+  MentorGradesTable,
+  StudentGradesTable,
+} from '@/features/courses/grades/course-grades-table.tsx'
+import CourseMainLayout from '@/features/courses/course-main-layout.tsx'
 
 export const mockStudentGradebook: CourseGradebookForStudent = {
   courseId: 'course_001',
@@ -326,285 +313,40 @@ const CourseGrades = () => {
   >([])
 
   return (
-    <Stack gap="md" h={'100%'} w={'100%'}>
-      <Group justify="space-between" align="start">
-        <Title>Grades</Title>
-      </Group>
+    <CourseMainLayout
+      title={'Grades'}
+      headerRightSection={
+        <Group align="start">
+          {role === 'student' ? (
+            <SearchComponent<StudentAssignmentGrade>
+              data={mockStudentGradebook.assignments}
+              onFilter={setStudentFiltered}
+              identifiers={['assignmentTitle']}
+              placeholder="Search..."
+            />
+          ) : (
+            <SearchComponent<CourseGradebookForMentor['assignments'][number]>
+              data={mockMentorGradebook.assignments}
+              onFilter={setMentorFiltered}
+              identifiers={['assignmentTitle', ['submissions', 'studentName']]}
+              placeholder="Search..."
+            />
+          )}
 
-      <Group justify="space-between" align="start">
-        {role === 'student' ? (
-          <SearchComponent<StudentAssignmentGrade>
-            data={mockStudentGradebook.assignments}
-            onFilter={setStudentFiltered}
-            identifiers={['assignmentTitle']}
-            placeholder="Search..."
-          />
-        ) : (
-          <SearchComponent<CourseGradebookForMentor['assignments'][number]>
-            data={mockMentorGradebook.assignments}
-            onFilter={setMentorFiltered}
-            identifiers={['assignmentTitle', ['submissions', 'studentName']]}
-            placeholder="Search..."
-          />
-        )}
-
-        <Group gap={rem(5)} justify="end" align="center">
-          <Button variant="default" radius="md">
-            Filters (to include)
-          </Button>
+          <Group gap={rem(5)} justify="end" align="center">
+            <Button variant="default" radius="md">
+              Filters (to include)
+            </Button>
+          </Group>
         </Group>
-      </Group>
-
+      }
+    >
       {role === 'student' ? (
         <StudentGradesTable assignments={studentFiltered} />
       ) : (
         <MentorGradesTable assignments={mentorFiltered} />
       )}
-    </Stack>
-  )
-}
-
-// --- Student View Table ---
-const StudentGradesTable = ({
-  assignments,
-}: {
-  assignments: StudentAssignmentGrade[]
-}) => (
-  <Box style={{ overflowX: 'auto', maxWidth: '100%' }}>
-    <Table
-      highlightOnHover
-      highlightOnHoverColor="gray.0"
-      style={{ borderRadius: rem('8px'), minWidth: '800px' }}
-    >
-      <Table.Thead>
-        <Table.Tr bg={'gray.1'} c={'dark.5'}>
-          <Table.Th>Assignment</Table.Th>
-          <Table.Th>Due</Table.Th>
-          <Table.Th>Submitted At</Table.Th>
-          <Table.Th>Status</Table.Th>
-          <Table.Th>Grade</Table.Th>
-          <Table.Th>Feedback</Table.Th>
-        </Table.Tr>
-      </Table.Thead>
-      <Table.Tbody>
-        <Suspense fallback={<SuspendedTableRows columns={6} />}>
-          {assignments.map((assignment) => {
-            // Get the latest submission
-            const latestSubmission =
-              assignment.submissions.length > 0
-                ? assignment.submissions[assignment.submissions.length - 1]
-                : null
-
-            // Determine if this is a group assignment
-            const isGroupAssignment =
-              latestSubmission?.grade?.groupId !== undefined
-
-            return (
-              <Table.Tr key={assignment.assignmentId}>
-                <Table.Td>
-                  <div>
-                    <Text fw={500}>{assignment.assignmentTitle}</Text>
-                    <Text size="sm" c="dimmed">
-                      {assignment.points} points
-                      {isGroupAssignment && ' (Group)'}
-                    </Text>
-                  </div>
-                </Table.Td>
-                <Table.Td>
-                  {formatTimestampToDateTimeText(assignment.dueDate, 'by')}
-                </Table.Td>
-                <Table.Td>
-                  {latestSubmission?.submissionTimestamp
-                    ? formatTimestampToDateTimeText(
-                        latestSubmission.submissionTimestamp,
-                      )
-                    : 'Not Submitted'}
-                  {latestSubmission?.isLate && (
-                    <Text size="xs" c="red">
-                      {latestSubmission.lateDays} day(s) late
-                    </Text>
-                  )}
-                </Table.Td>
-                <Table.Td>
-                  <Badge
-                    color={latestSubmission?.submissionStatus}
-                    variant="filled"
-                    size="sm"
-                  >
-                    {latestSubmission?.submissionStatus || 'pending'}
-                  </Badge>
-                  {assignment.submissions.length > 1 && (
-                    <Text size="xs" c="dimmed">
-                      {assignment.submissions.length} attempts
-                    </Text>
-                  )}
-                </Table.Td>
-                <Table.Td>
-                  {assignment.currentGrade ? (
-                    <div>
-                      <Text fw={500}>
-                        {assignment.currentGrade.score} /{' '}
-                        {assignment.currentGrade.maxScore}
-                      </Text>
-                      <Text size="xs" c="dimmed">
-                        {assignment.currentGrade.gradedAt &&
-                          `Graded ${formatTimestampToDateTimeText(assignment.currentGrade.gradedAt)}`}
-                      </Text>
-                    </div>
-                  ) : (
-                    <Text c="dimmed">- / {assignment.points}</Text>
-                  )}
-                </Table.Td>
-                <Table.Td>
-                  {assignment.currentGrade?.feedback ? (
-                    <Tooltip label={assignment.currentGrade.feedback}>
-                      <IconNote size={20} color="blue" />
-                    </Tooltip>
-                  ) : (
-                    <IconNote size={20} color="gray" />
-                  )}
-                </Table.Td>
-              </Table.Tr>
-            )
-          })}
-        </Suspense>
-      </Table.Tbody>
-    </Table>
-  </Box>
-)
-
-// --- Mentor View Table ---
-const MentorGradesTable = ({
-  assignments,
-}: {
-  assignments: CourseGradebookForMentor['assignments']
-}) => (
-  <Box style={{ overflowX: 'auto', maxWidth: '100%' }}>
-    <Table
-      highlightOnHover
-      highlightOnHoverColor="gray.0"
-      style={{ borderRadius: rem('8px'), minWidth: '1000px' }}
-    >
-      <Table.Thead>
-        <Table.Tr bg={'gray.1'} c={'dark.5'}>
-          <Table.Th>Assignment</Table.Th>
-          <Table.Th>Student</Table.Th>
-          <Table.Th>Status</Table.Th>
-          <Table.Th>Submitted At</Table.Th>
-          <Table.Th>Grade</Table.Th>
-          <Table.Th>Feedback</Table.Th>
-          <Table.Th>Actions</Table.Th>
-        </Table.Tr>
-      </Table.Thead>
-      <Table.Tbody>
-        <Suspense fallback={<SuspendedTableRows columns={7} />}>
-          {assignments.flatMap((assignment) =>
-            assignment.submissions.map((submission) => (
-              <Table.Tr
-                key={`${assignment.assignmentId}-${submission.studentId}`}
-              >
-                <Table.Td>
-                  <div>
-                    <Text fw={500}>{assignment.assignmentTitle}</Text>
-                    <Text size="sm" c="dimmed">
-                      {assignment.points} points • Due{' '}
-                      {formatTimestampToDateTimeText(assignment.dueDate, 'by')}
-                    </Text>
-                  </div>
-                </Table.Td>
-                <Table.Td>
-                  <Text>{submission.studentName}</Text>
-                  <Text size="xs" c="dimmed">
-                    ID: {submission.studentId}
-                  </Text>
-                </Table.Td>
-                <Table.Td>
-                  <Badge
-                    color={submission.submissionStatus}
-                    variant="filled"
-                    size="sm"
-                  >
-                    {submission.submissionStatus}
-                  </Badge>
-                </Table.Td>
-                <Table.Td>
-                  {submission.submissionTimestamp
-                    ? formatTimestampToDateTimeText(
-                        submission.submissionTimestamp,
-                      )
-                    : 'Not Submitted'}
-                </Table.Td>
-                <Table.Td>
-                  {submission.grade ? (
-                    <div>
-                      <Text fw={500}>
-                        {submission.grade.score} / {submission.grade.maxScore}
-                      </Text>
-                      <Text size="xs" c="dimmed">
-                        {submission.grade.gradedAt &&
-                          `Graded ${formatTimestampToDateTimeText(submission.grade.gradedAt)}`}
-                      </Text>
-                    </div>
-                  ) : (
-                    <Text c="dimmed">- / {assignment.points}</Text>
-                  )}
-                </Table.Td>
-                <Table.Td>
-                  {submission.grade?.feedback ? (
-                    <Tooltip label={submission.grade.feedback}>
-                      <IconNote size={20} color="blue" />
-                    </Tooltip>
-                  ) : (
-                    <IconNote size={20} color="gray" />
-                  )}
-                </Table.Td>
-                <Table.Td>
-                  <Button
-                    size="xs"
-                    variant="light"
-                    onClick={() => {
-                      // Handle grade/edit action
-                      console.log(
-                        'Grade assignment:',
-                        assignment.assignmentId,
-                        submission.studentId,
-                      )
-                    }}
-                  >
-                    {submission.submissionStatus === 'graded'
-                      ? 'Edit'
-                      : 'Grade'}
-                  </Button>
-                </Table.Td>
-              </Table.Tr>
-            )),
-          )}
-        </Suspense>
-      </Table.Tbody>
-    </Table>
-  </Box>
-)
-
-// --- Suspended Table Rows (Skeleton) ---
-const SuspendedTableRows = ({
-  columns = 6,
-  rows = 5,
-}: {
-  columns?: number
-  rows?: number
-}) => {
-  return (
-    <>
-      {Array.from({ length: rows }).map((_, rowIdx) => (
-        <Table.Tr key={rowIdx}>
-          {Array.from({ length: columns }).map((_, colIdx) => (
-            <Table.Td key={colIdx}>
-              <Skeleton height={20} radius="sm" />
-            </Table.Td>
-          ))}
-        </Table.Tr>
-      ))}
-    </>
+    </CourseMainLayout>
   )
 }
